Extract window related functions to a `Window` base class
Long term, I would expect `Window` to be not directly related to the
`Renderer` class, but instead allow creating a `Renderer` for a specific
viewport within the `Window`, such as for the window client area.

For now though, we just want to move things in the right direction.
Previously the window related methods were sorted together in the
`Renderer` class. Now they are moved out to a `Window` base class, so
the final `Renderer` interface is still basically the same. Eventually
the code may be updated so `Window` is no longer a base class, but
instead creates or returns a `Renderer` for a specific viewport.
