+++ NAS2D/Renderer/RendererNull.h
@@ -22,28 +22,6 @@ namespace NAS2D
 
 		~RendererNull() override {}
 
-		std::vector<DisplayDesc> getDisplayModes() const override { return {}; }
-		DisplayDesc getClosestMatchingDisplayMode(const DisplayDesc&) const override { return {}; }
-
-		void window_icon(const std::string&) override {}
-
-		void showSystemPointer(bool) override {}
-		void addCursor(const std::string&, int, int, int) override {}
-		void setCursor(int) override {}
-
-		void fullscreen(bool, bool = false) override {}
-		bool fullscreen() const override { return false; }
-
-		void resizeable(bool) override {}
-		bool resizeable() const override { return false; }
-
-		void minimumSize(Vector<int>) override {}
-
-		Vector<int> size() const override { return {}; }
-		void size(Vector<int>) override {}
-
-		Vector<int> getWindowClientArea() const noexcept override { return {}; }
-
 		void drawImage(const Image&, Point<float>, float = 1.0, Color = Color::Normal) override {}
 
 		void drawSubImage(const Image&, Point<float>, const Rectangle<float>&, Color = Color::Normal) override {}
